requirements.txt
django-oauth-toolkit
djangorestframework
drf-spectacular
requests
openpyxl
translators
python-dotenv